Refactoring components character-detail
