+++ src/pages/home-character-detail/character-detail.tsx
@@ -6,13 +6,15 @@ import {
     BtnBack,
     CharacterDetail,
     ITypeCharacterDetail,
+    ITypeEpisode,
     Loading,
     NotFound,
 } from "../../components";
-import { getCharacter } from "../../services";
+import { getCharacter, getEpisodePath } from "../../services";
 
 export const CharacterDetails = () => {
     const [character, setCharacter] = useState<ITypeCharacterDetail>();
+    const [episodes, setEpisodes] = useState<ITypeEpisode[]>([]);
 
     const [isLoad, setIsload] = useState(false);
     const [isNotFound, setIsNotFound] = useState(false);
@@ -25,6 +27,16 @@ export const CharacterDetails = () => {
                 setIsload(true);
                 const data = await getCharacter(id as string);
                 setCharacter(data);
+                const pathUrl = data.episode.map((url: string) =>
+                    url.slice(32),
+                );
+                const fetchEpisode = pathUrl.map(async (url: string) => {
+                    const dataEpisode = await getEpisodePath(url);
+                    return dataEpisode;
+                });
+                const allEpisodes: ITypeEpisode[] =
+                    await Promise.all(fetchEpisode);
+                setEpisodes(allEpisodes);
             } catch (error) {
                 // eslint-disable-next-line no-console
                 console.error(error);
@@ -41,7 +53,12 @@ export const CharacterDetails = () => {
             {(isLoad && <Loading />) || (isNotFound && <NotFound />) || (
                 <>
                     <BtnBack />
-                    {character && <CharacterDetail character={character} />}
+                    {character && (
+                        <CharacterDetail
+                            character={character}
+                            episodes={episodes}
+                        />
+                    )}
                 </>
             )}
         </Container>
